bitboard: initial bitboard
This commit adds support for uci commands "position->startpos->moves".
This is reflected in the bitboard which is primitive binary
representation of the board.

All pieces are stored as 64bit which represents each square on the
table.

With this commit, we can give the engine commands to move pieces:
 eg: a2a3->move piece from a2 to a3 which will result in handling the
piece at bit 8 and move it to bit 16.

Might add more information to this commit at one point.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>