bit_board: add knight moves
This commit adds initial knight moves.
The moves are generated at compile time so that the runtime will be an
O(1) look up.
The moves are then backtracked, also O(1), which should be quite a good
starting point.

The compile time code is not super pretty but it's evaluated at compile
time so it's probably fine for now.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>