kings_magic: add king move generation
This commit adds initial kings movement generation.
The moves are computed at compile time, similar to the knights table,
and is therefore a simple look up.

Also initial "king awareness" is added. Currently the solution is just
to move the king if under attack.
This ofc isn't the most efficient way to deal with a kings attack, but
we'll deal with that when we actually start adding value to our moves.

.. but for now, this should be a good starting point.

For this to work we also added a primitive for attack generation -
basically just a bit register of all squares that are currently under
attack.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>