movement: create namespace for piece movements
previously the movement were located under "magic" which is not entirely
true yet. We're gonna start looking into magic soon but until then we've
actually just generated the movements.

Created a sub folder for movements to keep things a bit more tidy.
Not the biggest fan of the word "movement", but I've failed to come up
with a better name. So I'll stick to this for now.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>