evaluator: add quiesence search
The engine would previously be tempted to sacrifice material without
actually knowing about it as it would search to a given depth and then
stop without knowing the consequences from that given move.

A common strategy to avoid this behavior is to add quiesence where you
perform an alpha beta pruning on all the potential captures in the given
position.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>