minors: fix some minor bugs + clean up some dead code
After introducing the MoveFlags I started seeing issues with castling.
Problem was that the uci_handler was not flag aware, so it would just
move pieces around without applying any flags. Not it is and special
moves are again working!

Also removed some dead code and cleaned up some helpers.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>