refactor: isolate bitboard from engine logic
The bitboard object was growing and becoming an issue to handle in
regards to dependencies.

Often we can to apply logic to the bitboard and not actually do engine
operations. Therefore I've separated the bitboard into its own object
and will instead use the new Engine object to handle move generation
etc.

This refactor will lead to more clean ups and easier implementation of
heuristics and fen parsing.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>