refactor: use bitboard/engine abstraction in move/atk generation
We can now improve the code given that we can safely pass the bitboard
with the generator methods.
This removes a lot of the ugly handling in the, now, engine class and
moves the logic to the actual generation - it can decide what to use
from the object itself without manipulating it.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>