scripts: add release and debug scripts
I'm mostly using two different builds; debug and release
To make it easier and better documented, I've added them as simple
scripts instead. I've decided to use bash as it's available on all Linux
distros :) or maybe because it's what I'm used to do in build scripts
and didn't want to bother too much. Anyways, it's a nice upgrade for my
personal workflow for now!

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>