heuristics: add triangular pv table
To get a better understanding of what's going on, I decided to implement
the PV table as the next heuristics.

The PV table will give us an idea about what the engine is thinking and
showing us the line it's predicting.

This also meant that we could do some clean up in the evaluator. This is
much better :)

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>