moves: 8byte align moves
When starting the project I was trying to optimize a bit too much on
readability and memory consumption.
Instead I've decided to memory align the moves to be optimized for
speed. This means that the CPU should have an easier time caching and
therefore improve performance.

Benchmarks: about 1 sec faster at perft 6

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>