perft: fix perft results
I was aware that my perft results were not perfect.
I did some debugging and now the perft results are correct.

It was a combination of wrong bitshift for black pawns' attacks.
Wrong side handling for check in the perft evaluator.
Wrong method called for black enpessant (white was called).

Also fixed some minors in relation to these fixes.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>