evaluator: add time control
Previously the engine would simply search to the preset depth.
As games can vary in length I've implemented some sort of time control
where we based on time inputs will consider how much time to spend and
try to search for as deep as we can.

This is quite primitive but it's better than before, so let's start with
that!

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>