engine: initial zobrist hash generation
This commit adds initial zobrist hashing.
More info can be found here:

https://en.wikipedia.org/wiki/Zobrist_hashing

The hashing tables are generated at compile time.
Currently the hash it not used but merely printed on debug to verify
that it's generated correctly.

NOTE: castling format does not currently support the hashing so I'll add
that later. Requires that the castling mask is formatted in a different
way.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>