tests: intial implementation of tests
Initial implementation of unit tests.
We're getting to a point where it's convenient to have some kind of
automated test. Breaking perft eg. is quite an indicator that something
went wrong.

I'm used to Catch2 so I added that as my testing framework.
I'll add more testing and automation later.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>