zobrist_hashing: add incremental hashing
To be able to use the zobrist hash, it's preffered to do what's known as
"incremental hashing". This means that we manipulate the hash on every
action. This is faster and has less overhead than computing a new hash
whenever we change the position.

For this to work properly a few helpers have been implemented as well.
BoardPosition eg. which will simplify a lot of the code going forward.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>