board_defs: Player enum instead of enum class
This commits changes our Player enum class to a simple enum.
The reason for this is to allow implicit casts.
We do this with other enums and is about time that we do the same here.

We need this for faster look ups that are player based (more to come :)
)

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>