bug: fix sorting corruption bug
std::sort (and std::stable_sort) requires strick comparators.
In our case, the move scores were different, as we were setting the
"isScoring" flag on first access to the PV mode and when it was found
another time it would be scored differently.
This would ultimatively lead to stack corruption as the search would be
inconsistent and therefore corrupted.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>