bug: fix aspiration window
I sometimes saw that my PV table was a very random. This could only mean
that my alpha and beta might have been reset.
It led to me investigating the aspiration window implementation and as I
expected; there was a bug. You should do a research with reset windows
BUT with same depth. Otherwise the PV will obvisouly be overwritten in
"bad positions" as 0 is better than nothing if eg. timed out.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>