tt_hash: add generation based aging
Instead of clearing the entire hash table on each position, we can
instead use a generation based aging mechanism. This means that we can
"outdate" a position if it's no longer relevant for us.
This will make our hash table waaay more useable as we'll be able to
reuse it between searches.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>