pv_table: increase capacity to avoid overflow
The PV table would previously have the same size as the max search
depth. This is problematic as the copy of moves will be into n+1.
Therefore we need an extra layer hence increasing the size of the tables
by one.

Without this the tables will overflow when reaching the max search
depth.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>