uci: add pondering
This commit adds pondering.

For this to work we had to start using multiple threads.
It was required to listen to the UCI while searching to know when to
start pondering etc.

This is more of an initial implementation. We'll move towards more multi
threading happening, so there'll be some generalization around threading
I assume. This is quite primitive but should work for now.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>