move_gen: refactor move to reduce redundant checks
In movegen we were doing a lot of "if white, else..". We can move this
to compile time and avoid all of these checks at all.

Also we can reduce the iteration overhead we introduced with
"getAllCaptures" and instead only populate the moves with captures from
the start. Again we can check for this at compiletime and let the
branching happen there.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>