meson: add ci build arch generalization
When building on the CI it would run with march=native. This meant that
the build was optimized for the server building it and not in general
optimized for most x86 cpus.

Using the x86-v2 we can build for almost every CPU out there. And it
should also be optimized way further.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>