meson: add cross compilations
This commit adds a very flexible way of adding multiple build targets.
As we'll have some features available based on CPU arch, it's very quite
practical to be able to distinguish from these builds.

Also in this way, we can create a very optimized build for when building
locally (arch=native) and otherwise a more generic build based on x86
arch (for now).

This also means, that we should easily be able to add windows, ARM etc.
targets if needed at one point.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>