evaluation: add fundamentals for game phases evaluation
This commit adds a way to generalize our scorings.
Instead of iterating multiple times, all scorings are done in the single
piece iteration.

Also it means that all our scores are now tapered. This will add more
parameters for us to tune ie. more options to make the engine stronger!

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>