tests: add cutechess tournament tests
This commit adds a new way of testing the engine.
It will run games with a given config and evaluate the performance
between the two engines.

It is possible to setup the amount of games, if the test should run SPRT
etc.
More info is added to the README.md.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>