tests: add script to build baseline
This commit adds a helper script to build a baseline for testing.
You can simply call this script with the branch you want to build from.
The script will ensure your current git state is restored etc.

Reasoning: I found myself building main as a baseline and it becamse
annoying to handle that myself. Now I can just call this script and it's
automatic.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>