uci_handler: add authors print out
This commit adds auto generated author information.
I can't take credits for all the engine work, so I want to make sure
that authors that help contributing to the engine is also praised. Their
name will now be embedded in the engine as a way to show appreciation
for the support and contributions :)

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>