terms: correct pawn shield scores
Previously the shield was calculated as any piece in the attack mask of
the king. This is not necessarily a good indicator for shielding. The
queen could eg. be the shield.

Instead turn the score into a pawn shield: pawns in front of the king
where the score is given based on how far the pawn is from the king.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>