uci_handler: add uci options generalization
Previously it was quite a pain to add new UCI options.
You had to add it manually to the "uci ok string", the debug handle and
manually parse all the data.

Now instead, all you have to do is add it to the list and implement the
callback.

Adding new types is also easy but requires a bit more work.

This implementation should be better, safer and way faster for new
additions.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>