uci: add Hash option
Previously the TT hash table was a fixed size.
This commit adds dynamic scaling of the table.

The option has been added to the UCI handler so it can be provided as an
UCI option.

Due to the atomics constructors I decided to use our own
allocator. It seemed simpler than trying to make it work with a std
container.

Bench 13605379

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>