engine: add bench
This commit adds the option to "bench" the engine.
This is a "more or less standardized" way of benchmarking engines.

The positions are used by most engines out there.

We have two ways of running a bench:
From the UCI handler: bench <depth>
From cmd line directly: ./meltdown bench

The latter is necessary for eg. openbench.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>