makefile: add openbench complicant makefile
This commit adds support for building meltdown using a makefile.
This is necessary to support OpenBench.

The makefile should not be used instead of the provided build scripts.
If a build is triggered using the makefile it will "check" if it's an
OpenBench instance. Otherwise the build is rejected.

To support OpenBench we need to append the node count as following:
Bench 13599361

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>