meson: reduce cpp version to cpp23
When I started this project I decided to use the newest available
compiler - just for the heck of it.
Now that the project has grown a bit, it's become clear that it's not
necessarily a benefit over a slightly older one.

(btw) I'm using Arch meaning that I have access to the newest compiler
etc, but when setting up automated builds, server instances etc, they
often run Ubuntu which doesn't have access to the same bleeding edge
versions.

As we don't currently use any cpp26 specific features, I think the
simplest is to reduce the cpp version to 23 meaning that it will now be
easier to build for people not using Arch etc.

Bench 13598897

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>