lazy_smp: fix early-return bug
Upon a full search being completed by another thread, the current
searcher thread was always returning alpha for its current node. This
was a bug when the node is a minimizer, and if we're very unlucky it
could theoretically return -infinity for a node's score.

The early termination now follows the `stop` termination, i.e. each node
return a value from its tree, even if that value is suboptimal.