tt_hash_table: remove generation
When implementing the TT hash table I introduced a so called
"generation" to the entries.
What this meant is that any search made from a previous search (when
playing a continous game) would be invalidated.
But why should it be invalidated? We already searched that position. It
doesn't change? We still compare the score to our current beta and alpha
values, so it should be safe to store the position and reuse it if
needed.

.. this is basically the essence of having a hash table, sigh.

Bench 14657082

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>