tt_hash_table: move alpha beta comparison to consumer
As we introduced TT move ordering there's no reason to reject a probe
just because it's not a cutoff. It's actually a shame to throw that data
away when we can use it for move ordering..

This commit ensures that the probe is always return and will now contain
the entry flag as well. Then the consumer can decide what to do with the
data.

In our case: return early during cutoffs or exact score - or use the
score as the TT move for move ordering.

Bench 13322160

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>