workflows: add verify_commit_bench to pull_request workflow
This commit adds a check for pull request that compares the bench result
on the compiled engine with the one provided in the commit message.

If it's missing an error will print the expected bench result and how to
avoid the error.
If the bench value is not updated then the new bench value will be
provided as well.

NOTE: if multiple bench results are added, or if you include Bench in
the commit message, without specifying a value, then it should still
work (yes this is a test..). The last provided bench value will be used.
So in our case:
Bench 123456
will not cause an error (yes, another test..)

Final.. add the actual bench value..
Bench 13598897

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>