terms: add virtual mobility
This commit adds the term "virtual mobility".

This term is fairly simple: we replace the king with a queen (in theory)
and calculates the mobility if the king was a queen. This will give us
an idea of how the king is exposed to sliding pieces which might put the
king in danger.

Bench 11029400

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>