thread_pool: implement thread pool and add to evaluator
This commit adds a thread pool.
The thread pool is designed to be fast, simple and safe to use.
The threads' lifetime is ensured in the thread pool's life time - ie. no
way to kill threads in an unsafe way.

There is one implementation detail that is necessary: scaling the amount
of workers at runtime. Therefore the thread pool has a safe way to kill
the threads and respawn new ones based on the required runtime size.

The resize method is blocking and will only work if the threads are
allowed to terminate.

A simple LIFO queue has been added as well.
The reason for LIFO is to keep the pool simple and fast. For our usecase
the LIFO should be totally fine.

To use the thread pool in the Evaluator, I've added a new handle:
"getBestScoreAsync". This method will do the same as the non-async
variant, but instead it will be done... async.
The reason for this is to allow the consumer to have a simple handle to
the evaluator where all threading is handled there. No need to worry
about offloading etc. -it's not the job of the consumer - it's for the
owner of the threads.

Bench 11349788

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>