hack: fix time handler's timeout handling
When running the bench it's swiching between runs fairly fast.
The time handler is only checking for updates every 1ms. This opens up
for a race condition where the time manager might occupy a thread and
also not set the "isStopped" state.

The time handler is gonna get nuked soon, so this is just a hack to test
this branch.

Bench 11349788

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>