tests: auto build engines for tournaments
This commit adds the option to automatically build engines for the
tournament test.

An optional argument has been added to the script.
If a branch is provided, it will be built as a baseline for the test.
The dev engine will be the currently checked out branch.

Room for improvement here but it should improve local testing a fair bit
:)

Bench 11349788

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>