tests: simplify build engines
This commit makes the build engines helper a bit more verbose and also
more fail proof.
Before the stash would be auto applied etc. but that's maybe a little
dangerous.
Instead let the user handle that.

Also now the build is a bit more verbose and fail proof.
The cleanup ensures that we always return to the currently checked out
branch.

Bench 11349788

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>