time_manager: fix printing on stop()
New time manager added a bug where upon calling stop(), an extra
`printScoreInfo` call happened, printing an empty PV line.

This happened when `stop` command was entered mid-search. Search
terminates and PV is not updated. But it gets printed anyway.

This commit wraps printing with an additional check for the stopped
condition, fixing the bug.

Bench 11349788