time_manager: add soft and hard limits
Our current time management is absolute... sheit :) .. It was just
something initial that had to be fixed at some point.

Sam Roelants was so kind to write a paper about the topic:
https://github.com/sroelants/simbelmyne/blob/main/docs/time-management.md

So this commit adds a better foundation for our time management.
This is by no means final, but a way better starting point than what we
already have. And it's a whole lot simpler..

Bench 11349788

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>