evaluation: add generic 16bit score type
Currently the score is stored as int32_t type.
The reason for this was the score range which would go from [-50k:50k].
But as we control the range ourselves, we might as well just scale down
the scores to fit in 16bit type as well.

This allows us to simplify our TTHashEntryData to not embed partial
scores etc. Now it's simple, clear and the values are smaller -> fits
better in most chess GUIs.

NOTE: the new "Score" type has been made explicit for a few reasons:
1. I want to be explicit about the value is a score. Obviously
2. It's easier to follow what's a score. Obviously
3. I want to add helpers to our score handling and those will go in the
   Score.h file. Hence its own file.

Bench 11029400

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>