tt_hash_table: improve hash full
Previously we would keep incrementing the counter whenever we wrote an
entry. This was merely an approximation of how full the table was.

Another, and more efficient, way is to take a random snapshot of the
table and check each of those and extract a permill.

The simplest, and most efficient of above, is to take a sample base of
1000 entries as we'll already have the permill as result. Easy :)

Bench 11029400

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>