tt_hash_table: make read and write non-atomic
We're using relaxed memory access in all our atomic calls so there's not
actually any need to do these operations atomically.

For now, this is not a big difference, but as we can pack our entries
even closer together - and add more values to the entry - this becomes
more prominent.

This means, as before, that all read and write operations are racy. But
as we always perform thread voting this will be a sanity check anyways
hence it doesn't matter too much.
Syncronizing the threads would lose more elo than the potential gains.

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>