tt_hash_table: improve read and write filtering
Prior to this commit we would have some issues regarding read and write
of TT entries.

Let's start with read:
We would only read an entry if it was from a deeper search than our
current. This makes sense for our entry test - only return if it's
actually an exact entry or will cause a cut off.
However, as we also use the probe for move ordering, we're actually
losing out on a lot of ttMoves that would otherwise give us an edge.

And write..
Well.. we had no filter meaning that any result could overwrite any
entry. So we could have spent forever searching a branch, store it in
the TT, just to overwrite it with a lesser search. Not very efficient.
Now we only overwrite the TT entries if there's a better one found.

Bench 8880290

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>