searcher: add internal iterative reductions
Internal Iterative Reductions (IIR) is a newer and more commonly used
variant of Internal Iterative Deepening (IID).

The idea is that when no hash move is found, the guess is that the
node must not be very important (at the current iteration).

Instead search one shallower depth and let the next iteration handle the
deeper level (if found more important at this point).

Bench 8385565

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>