searcher: add search stack
This commit adds a search stack for the searchers.
Previously we stored a local hash copy + "new board" when performing the
search recursion.

It is beneficial to actually store all the positions we've been through
during the search. This means that we can, in theory, iterate the stack
and get a better understanding on how the current search is performing.

Furthermore it should improve the cache as we no longer allocate the
temporary hash and board on the stack.

Bench 8741465

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>