searcher: add isImproving heuristics
This commit takes advantage of the previously implemented search stack.
We can now compare our iterations and decide if our search is improving.
If so, we can be a little more aggresive in our pruning.

The first example is null move pruning; here we can perform an even more
aggressive pruning if we know that we're on the "safe side" of a search.

Bench 6859335

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>