move_picker: create move picker
The current sorting of moves is very wasteful: some moves won't be used
due to pruning but we generate them anyway.

This PR replaces the move ordering with a phased move picker, that
attempts to first pick a TT move, then a PV move, good capture, ...

This commit is for the basic picker. There are plenty of optimizations
to be had afterwards.

Bench 3895729