tt: add evaluation to entries
As our static evaluation grows in complexity it also grows in time to
compute. Therefore it might be beneficial to add the evaluation to our
last slot in the transpostion table.

This commit adds the possibility to add a lighter entry in our tables.
It will contain no score nor move. Therefore I've also added a term
"noScore" and created a helper to create null moves.

As these "entries" are solely used for evaluation we ofc allow a full
search to be stored over a simple evaluation.

Bench 8688357

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>