tuner: add time estimates
This is totally not necessary, but it is cool nevertheless.
This commit adds a visual representation of how long the tuning has been
running and an estimate on how long the tuning is gonna run.

It's nice if you want to go juggling, grab a cup of coffee etc. :)

Bench 7987248

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>