tuner: improve learning rate
Previously we would run 2000 epochs and have a gradually reduced
learning rate. This was merely a starting point to get things going with
the tuner.

To squeeze out as much performance of the tuning as possible we should
first of all perform more iterations. Therefore I've increased it to
2500 iterations. This is a slight increase, for now, but it's when I
noticed that the MSE didn't change for the last 100~ epochs.

A problem with running a higher number of iterations is that we might
end up in a situation where we start over-fitting. This is especially
the case with lower values learning rates (which we would use in the end
due to learning rate adjustments).

BUT, there is one detail in our implementation; we're using an adaptive
gradient descent algorithm; adamW. So there's actually no need for
adjusting our learning rate manually. Therefore the learning rate
adjusments have been removed and we'll now only rely on the adamW
parameters.

Bench 8564539