terms: add pawn attacks to mobility scores
This commit utilizes the newly added "TermContext". We can now bring
pawn attacks into other terms with literally no cost.

This means that we can improve our mobility terms; only apply the score
if the square is not attacked by opponent's pawns.

Also it fixes a minor bug where we included our own pieces' squares in
the mobility. This is of course not valid, so we should not include
that.

Bench 4065417

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>