searcher: add lmr table
This commit adds dynamic scaling of the Late Move reductions.
Previously we had a fixed reduction (2) which was a fine starting point.

With this commit we scale the reduction based on two parameters: depth
and amount of moves already searched.

This means that the deeper we're into a search and the more move's we've
searched at that depth the higher reduction.

This is quite huge which is also reflected in the bench nodes. We're
pruning a lot now!

The table values are borrowed from the engine Carp - a lot of engines
use these values these days, so it's probably not a bad starting point
:)

Bench 4397450

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>