searcher: add zugzwang check to null move pruning
A known issue with null-move pruning is that if performed on a zugzwang
position it's actually allowing you to improve your own position. This
is not exactly what we expect from null-move pruning, so let's add a
check where we approximate the zugzwang position.

Zugzwang is kinda rare, but it's most likely to happen in endgames where
you only have king and pawns left on the board. So this should be a fine
approximate without losing too much time computing rare positions.

Bench 8562503

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>