evaluator: correct tbhits print out
Previously the TbHits would always print either 0 or 1. It's not
actually necessary to print the value if there's no hit, so first of
all, avoid doing that.

Second, we should actually print the amount of table hits and not just
if we hit a table. So this commit also changes that.

Bench 4297319

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>