time_manager: add uci option for move overhead
When using the engine online, eg. lichess bot, I've noticed a fair
amount of time outs in games that would otherwise be won.

Our previous 50ms move overhead is not much, taking network traffic into
account. Therefore allow setting the move overhead from the UCI handler,
so we can correct for this.

Also with the new and improved way of handling the move overhead, this
is finally safe to do.

Bench 4297319

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>