searcher: improving heuristics - add to lmr
This commit adds initial "improving heuristics".

Improving heuristics is essentially comparing our static eval with our
previous static eval (current_ply - 2) and see if it's improved since
then.

If that's the case we can adjust our pruning.
Eg for LMR we can increase the reduction when not improving as this line
is not as interesting as if we were improving.

This pattern can be added to more pruning, but for now, only add it to
LMR.

NOTE: I tried also decreasing the reduction when we're improving but it
was not as efficient as not doing so.

Bench 3161528

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>