searcher: correct is stopped handling
Previously we would return the current score when stopped. This is an
issue as the score has not yet been accounted for in regards to alpha
and beta. Ie. this value was assumed to always be alpha.

Instead treat the score as invalid, so ensure we don't use the result. Hence
returning s_minScore -> will always fall outside the aspiration window.

This also fixes a long term bug where we print a PV line that is not
actually fully searched.

This also means that we don't store an invalid score in the TT which
could potentially be returned without checking our stop conditions (if
stored right after root move).

Bench 4016654

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>