searcher: correct syzygy threading issue
I saw a few end-games getting stuck on Lichess.
The `tb_probe_root` is not thread safe so let's use the newly added
"isPrimary" to ensure only a single thread will take this path.

This is a simple and efficient way to ensure we still have the same
results during single threaded search and also still gain the full power
of the syzygy probe.

Bench 4016654

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>