terms: add un/safe checks terms
This commit adds two new terms:

Unsafe and safe checks.

A safe check is considered when a piece can move into a position where
it can check the opponent's king without being attack.

An unsafe check is when a piece can move into a position where it can
check the opponent's king while being attacked.

We have most of the handles provided already, so it should be a small
runtime cost.

Bench 4016654

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>