transposition: add tt pv
This commit adds PV logic to the TT entries.
This means that the we allow earlier writes if PV (we assume it's a more
imporant entry) and thereby have a way to track the PV line.

As the TT move is the first picked, this will result in PV move ordering
to be redundant.

Bench 1610421

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>