searcher: let all moves be considered lmr reductable
We're applying reductions based on the move, board state etc. so let all
moves be candidates for lmr reduction.
Also with our "new" lmr table we can be more confident about which moves
we allow reduction as the table is already reflecting the search state
(depth and moves searched).

Bench 2797095

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>