searcher: make beta static eval aware
Our previous qsearch implementation was rather stupid.
It would just play through all captures and check for static evals.

Problem is that we can not just count on static eval. The board can be a
draw, check mate etc. So in case of draws we might actually think that
the board is winning when it's in fact not.

This commit refactors how we handle beta cut offs - they need to be
propagated all the way from qsearch to root.

Bench 2338571

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>