searcher: add improving heuristics
This commit adds initial "improving heuristics".

The improving heuristic is a scalar that can be appended to various
other heuristics. Ie search less deep when not improving (not an
important positition?) and maybe search deeper when improving (more
important position?).

As a start simply append the improving heuristic to LMR where we can
reduce the search depth if the position is not improving.

Bench 1678321

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>