build: default to llvm toolchain
The LLVM toolchain seems to be the golden standard for chess engines.
I've conducted a few tests and it's literally insane how much of a
difference it makes!

Let's default to clang going forward!

NOTE: I've kept the GCC targets as they might come handy for people
trying to compile the engine natively etc.

Bench 2955703

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>