terms: add candidate passers
This commit adds the term "candidate passers".
This term is divided into two parts; defended and undefended.

A candidate passer is when we are a pawn up in case of eg. trading pawns
and the trade will result in a passed pawn.
We check this by comparing threats and support for the current pawn's
position and the future position it will enter when pushed.

Bench 1566248

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>