static_evaluation: add king pawn cache
As king pawn evaluation is more or less static we can store this "static" evaluation in a cache.
This will give us a fair amount of speed up as we'll avoid computing a
fair amount of duplicate work.

NOTE: this only works for king-pawn as there's not that many options
from a given position.

One downside: we also have to be careful in the future when adding new terms as we
need to beaware of them being static or not.

Bench 1555918

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>