search: add initial correction history
This commit adds a base for Correction History.

Initially the kpHash is used for a simple and fast correction.
The correction is applied on the static eval as an attempt to average
out evaluation errors for positions where we have eg. same kp structure.

This means that we ultimatively will have a more precise static
evaluation, as we can mitigate potential evaluation errors over time.

Bench 2923161

Signed-off-by: Hans Binderup <hbinderup94@gmail.com>